requirements.txt
tensorflow
numpy
pillow
streamlit
matplotlib
scikit-learn
